+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/ItLivenessProbeCustomization.java
@@ -26,14 +26,12 @@ import oracle.weblogic.domain.ServerPod;
 import oracle.weblogic.kubernetes.annotations.IntegrationTest;
 import oracle.weblogic.kubernetes.annotations.Namespaces;
 import oracle.weblogic.kubernetes.logging.LoggingFacade;
-import org.awaitility.core.ConditionFactory;
+import oracle.weblogic.kubernetes.utils.ExecResult;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
-import static java.util.concurrent.TimeUnit.MINUTES;
-import static java.util.concurrent.TimeUnit.SECONDS;
 import static oracle.weblogic.kubernetes.TestConstants.ADMIN_PASSWORD_DEFAULT;
 import static oracle.weblogic.kubernetes.TestConstants.ADMIN_SERVER_NAME_BASE;
 import static oracle.weblogic.kubernetes.TestConstants.ADMIN_USERNAME_DEFAULT;
@@ -43,21 +41,23 @@ import static oracle.weblogic.kubernetes.TestConstants.MII_BASIC_APP_NAME;
 import static oracle.weblogic.kubernetes.TestConstants.TEST_IMAGES_REPO_SECRET_NAME;
 import static oracle.weblogic.kubernetes.TestConstants.WLS_DOMAIN_TYPE;
 import static oracle.weblogic.kubernetes.actions.ActionConstants.RESOURCE_DIR;
+import static oracle.weblogic.kubernetes.actions.TestActions.execCommand;
 import static oracle.weblogic.kubernetes.actions.TestActions.getContainerRestartCount;
 import static oracle.weblogic.kubernetes.actions.TestActions.getDomainCustomResource;
 import static oracle.weblogic.kubernetes.actions.impl.primitive.Kubernetes.copyFileToPod;
 import static oracle.weblogic.kubernetes.assertions.TestAssertions.appAccessibleInPod;
 import static oracle.weblogic.kubernetes.assertions.TestAssertions.appNotAccessibleInPod;
 import static oracle.weblogic.kubernetes.utils.CommonTestUtils.checkPodReadyAndServiceExists;
+import static oracle.weblogic.kubernetes.utils.CommonTestUtils.testUntil;
 import static oracle.weblogic.kubernetes.utils.DomainUtils.createDomainAndVerify;
 import static oracle.weblogic.kubernetes.utils.ImageUtils.createMiiImageAndVerify;
 import static oracle.weblogic.kubernetes.utils.ImageUtils.createTestRepoSecret;
 import static oracle.weblogic.kubernetes.utils.ImageUtils.dockerLoginAndPushImageToRegistry;
 import static oracle.weblogic.kubernetes.utils.OperatorUtils.installAndVerifyOperator;
+import static oracle.weblogic.kubernetes.utils.PodUtils.checkPodReady;
 import static oracle.weblogic.kubernetes.utils.PodUtils.setPodAntiAffinity;
 import static oracle.weblogic.kubernetes.utils.SecretUtils.createSecretWithUsernamePassword;
 import static oracle.weblogic.kubernetes.utils.ThreadSafeLogger.getLogger;
-import static org.awaitility.Awaitility.with;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -65,8 +65,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 /**
  * Test liveness probe customization in a multicluster mii domain.
  * Build model in image with liveness probe custom script named 
- * customLivenessProbe.sh that retuns success(0) when a file /u01/tempFile.txt
- * avilable on the pod else it returns failure(1). 
+ * customLivenessProbe.sh that returns success(0) when a file /u01/tempFile.txt
+ * available on the pod else it returns failure(1).
  * Note: Livenessprobe is triggered only when the script returns success 
  */
 
@@ -91,11 +91,6 @@ class ItLivenessProbeCustomization {
 
   private static LoggingFacade logger = null;
 
-  private static ConditionFactory withStandardRetryPolicy =
-      with().pollDelay(2, SECONDS)
-          .and().with().pollInterval(10, SECONDS)
-          .atMost(5, MINUTES).await();
-
   /**
    * Get namespaces for operator and WebLogic domain.
    *
@@ -118,7 +113,7 @@ class ItLivenessProbeCustomization {
     // install and verify operator
     installAndVerifyOperator(opNamespace, domainNamespace);
 
-    // create mii with additional livenessprobecustom script
+    // create mii with additional livenessprobe custom script
     String imageName = createAndVerifyDomainImage();
 
     // create a basic model in image domain
@@ -242,6 +237,95 @@ class ItLivenessProbeCustomization {
   }
 
   /**
+   * Verify liveness probe by killing managed server process 3 times to kick pod container auto-restart.
+   */
+  @Test
+  @DisplayName("Test liveness probe of pod")
+  void testLivenessProbe() {
+    Domain domain = assertDoesNotThrow(() -> getDomainCustomResource(domainUid, domainNamespace),
+        String.format("getDomainCustomResource failed with ApiException when tried to get domain %s in namespace %s",
+            domainUid, domainNamespace));
+    assertNotNull(domain, "Got null domain resource");
+
+    String domainUid = domain.getSpec().getDomainUid();
+    String domainNamespace = domain.getMetadata().getNamespace();
+
+    String serverNamePrefix = domainUid + "-cluster-1-" + MANAGED_SERVER_NAME_BASE;
+
+    // create file to kill server process
+    File killServerScript = assertDoesNotThrow(() -> createScriptToKillServer(),
+        "Failed to create script to kill server");
+    logger.info("File/script created to kill server {0}", killServerScript);
+
+    String server1Name = serverNamePrefix + "1";
+    checkPodReady(server1Name, domainUid, domainNamespace);
+
+    // copy script to pod
+    String destLocation = "/u01/killserver.sh";
+    assertDoesNotThrow(() -> copyFileToPod(domainNamespace, server1Name, "weblogic-server",
+        killServerScript.toPath(), Paths.get(destLocation)),
+        String.format("Failed to copy file %s to pod %s in namespace %s",
+            killServerScript, server1Name, domainNamespace));
+    logger.info("File copied to Pod {0} in namespace {1}", server1Name, domainNamespace);
+
+    // get the restart count of the container in pod before liveness probe restarts
+    final int beforeRestartCount =
+        assertDoesNotThrow(() -> getContainerRestartCount(domainNamespace, null, server1Name, null),
+            String.format("Failed to get the restart count of the container from pod {0} in namespace {1}",
+                server1Name, domainNamespace));
+    logger.info("Restart count before liveness probe {0}", beforeRestartCount);
+
+    // change file permissions
+    ExecResult execResult = assertDoesNotThrow(() -> execCommand(domainNamespace, server1Name, null,
+        true, "/bin/sh", "-c", "chmod +x " + destLocation),
+        String.format("Failed to change permissions for file %s in pod %s", destLocation, server1Name));
+    assertTrue(execResult.exitValue() == 0,
+        String.format("Failed to change file %s permissions, stderr %s stdout %s", destLocation,
+            execResult.stderr(), execResult.stdout()));
+    logger.info("File permissions changed inside pod");
+
+    /* First, kill the managed server process in the container three times to cause the node manager to
+     * mark the server 'failed not restartable'. This in turn is detected by the liveness probe, which
+     * initiates a container restart.
+     */
+    for (int i = 0; i < 3; i++) {
+      execResult = assertDoesNotThrow(() -> execCommand(domainNamespace, server1Name, null,
+          true, "/bin/sh", "-c", destLocation + " " + server1Name),
+          String.format("Failed to execute script %s in pod %s namespace %s", destLocation,
+              server1Name, domainNamespace));
+      logger.info("Command executed to kill server inside pod, exit value {0}, stdout {1}, stderr {2}",
+          execResult.exitValue(), execResult.stdout(), execResult.stderr());
+
+      try {
+        Thread.sleep(2 * 1000);
+      } catch (InterruptedException ie) {
+        // ignore
+      }
+    }
+
+    // check pod is ready
+    checkPodReady(server1Name, domainUid, domainNamespace);
+
+    // get the restart count of the container in pod after liveness probe restarts
+    int afterRestartCount = assertDoesNotThrow(() ->
+            getContainerRestartCount(domainNamespace, null, server1Name, null),
+        String.format("Failed to get the restart count of the container from pod {0} in namespace {1}",
+            server1Name, domainNamespace));
+    assertTrue(afterRestartCount - beforeRestartCount == 1,
+        String.format("Liveness probe did not start the container in pod %s in namespace %s",
+            server1Name, domainNamespace));
+
+    for (int j = 1; j <= replicaCount; j++) {
+      String managedServerPodName = domainUid + "-cluster-1-" + MANAGED_SERVER_NAME_BASE + j;
+      String expectedStr = "Hello World, you have reached server " + "cluster-1-" + MANAGED_SERVER_NAME_BASE + j;
+      checkAppIsRunning(
+          domainNamespace,
+          managedServerPodName,
+          expectedStr);
+    }
+  }
+
+  /**
    * Create a model in image domain and verify the server pods are ready.
    */
   private static void createAndVerifyMiiDomain(String miiImage) {
@@ -359,22 +443,12 @@ class ItLivenessProbeCustomization {
   ) {
 
     // check that the application is NOT running inside of a server pod
-    withStandardRetryPolicy
-        .conditionEvaluationListener(
-            condition -> logger.info("Checking if application {0} IS running on pod {1} in namespace {2} "
-            + "(elapsed time {3}ms, remaining time {4}ms)",
-            appPath,
-            podName,
-            namespace,
-            condition.getElapsedTimeInMS(),
-            condition.getRemainingTimeInMS()))
-        .until(() -> appAccessibleInPod(
-                namespace,
-                podName,
-                internalPort,
-                appPath,
-                expectedStr));
-
+    testUntil(() -> appAccessibleInPod(namespace, podName, internalPort, appPath, expectedStr),
+        logger,
+        "application {0} is running on pod {1} in namespace {2}",
+        appPath,
+        podName,
+        namespace);
   }
 
   private static void checkAppNotRunning(
@@ -384,21 +458,12 @@ class ItLivenessProbeCustomization {
   ) {
 
     // check that the application is NOT running inside of a server pod
-    withStandardRetryPolicy
-        .conditionEvaluationListener(
-            condition -> logger.info("Checking if application {0} is NOT running on pod {1} in namespace {2} "
-            + "(elapsed time {3}ms, remaining time {4}ms)",
-            appPath,
-            podName,
-            namespace,
-            condition.getElapsedTimeInMS(),
-            condition.getRemainingTimeInMS()))
-        .until(() -> appNotAccessibleInPod(
-                namespace,
-                podName,
-                internalPort,
-                appPath,
-                expectedStr));
+    testUntil(() -> appNotAccessibleInPod(namespace, podName, internalPort, appPath, expectedStr),
+        logger,
+        "app {0} is not running on pod {1} in namespace {2}",
+        appPath,
+        podName,
+        namespace);
   }
 
   private static String createAndVerifyDomainImage() {
@@ -440,5 +505,23 @@ class ItLivenessProbeCustomization {
     return tempFile;
   }
 
-
+  /**
+   * Create a script to kill server.
+   * @return a File object
+   * @throws IOException if can not create a file
+   */
+  private File createScriptToKillServer() throws IOException {
+    File killServerScript = File.createTempFile("killserver", ".sh");
+    //deletes the file when VM terminates
+    killServerScript.deleteOnExit();
+    try (FileWriter fw = new FileWriter(killServerScript)) {
+      fw.write("#!/bin/bash\n");
+      fw.write("jps\n");
+      fw.write("jps | grep Server\n");
+      fw.write("jps | grep Server | awk '{print $1}'\n");
+      fw.write("kill -9 `jps | grep Server | awk '{print $1}'`");
+    }
+    killServerScript.setExecutable(true, false);
+    return killServerScript;
+  }
 }
